Remove newsletter form
The newsletter form is a great tool, but I do not have the time to keep
it going as well as I would like. For now, people can follow me on
Twitter and Facebook to keep up with my posts. Some will be at Linkedin
too.

Doing something is better than not doing, but as I promise "great
reading", I can't send some halfly baked reading list. Maybe if I go
freelance or after the college.